convert_to_pdf: Remove dependencty on pandoc-latex-environment
The core reason for including the pandoc-latex-environment was to be
able to use the custom-defined `questionblock` environment for wrapping
questions.

However, the usage of pandoc-latex-environment comes with additional
whitespace being produced that causes formatting issues. Since we have
to include code to generate ::: fences surrounding the question text
anyway, we might as well generate the raw latex envinronment directly.