utils: Became robust against random foldclosing errors
For a weird reason, although we are performing the fold close operation
on the correct line (and with a check to perform it only if the fold is
not closed already!) neovim complains on startup that the fold is
already closed.

On the other hand, if the fold-close procedure is completely turned off
via the knowledge_autoclose_question setting, the closing procedure does
not happen.

There this commit opts to supress this error instead.