IntegrationTest: Generate expected tag from the deck name (Mnemosyne only)
In MnemosyneProxy, the deck information is being stored as a
(potentially nested) tag. This is due to the fact that facts in Mnemosyne
do not have an explicit "deck" attribute.

The tests were previously hardcoded to expect the deck name in the
Mnemosyne format in the list of tags, but now that we are moving to
supporting both Anki and Mnemosyne in tests, that is being adjusted.