ci: Temporarily disable docker image layer caching
Currently, docker builds with --cache-from cause testing failures due to
creation of 0B-sized layers, which manifests as i.e. xvfb-run missing
in the test image. This is a known issue [1] and fixed in 19.03.13.

However, we have no way of upgrading the version of the docker ourselves
and have to wait for the AKS team to upgrade to the newest stable
release (19.03.13 was just released a week ago).

This commit should be reverted once docker version is upgraded to
19.03.13.