latex: Make occlusions behave inversly to other OCG groups
Since image occlusions are overlaying content, they should be visible
when the content to remember is in the hidden state, and hidden when the
content to remember is in the visible state.

This behaviour is inverse to how regular hidden OCGs groups work, so we
need to define a new container \allocclusions and store the OCG
references for occlusions there.